split up DET stuff into a standalone package
